pyproject.toml
[project]
name = "comfyui-image-process"
version = "1.0.00"
description = "A ComfyUI custom node collection for image processing."
authors = [
  {name = "Wakaura Asaho", email = "onestandaction@gmail.com"}
]
readme = "README.md"
license = { file = "LICENSE" }
classifiers = []
dependencies = [
  "comfyui-frontend-package>=1.37.11",
  "numpy>=2.3.5",
  "Pillow>=12.1.0",
  "scipy>=1.16.3"
]

[project.urls]
bugs = "https://github.com/wakaura-asaho/comfyui-image-process/issues"
homepage = "https://github.com/wakaura-asaho/comfyui-image-process"
Repository = "https://github.com/wakaura-asaho/comfyui-image-process"

[tool.comfy]
PublisherId = "wakaura"
DisplayName = "ComfyUI-Image-Process"
Icon = ""
requires-comfyui = ">=0.12.3"